Major overhaul, now everything looks like its set to proceed

+++ get_errno.h
@@ -1,6 +1,7 @@
 #include <errno.h>
+#include "u.h"
 
 #ifndef GET_ERRNO_H
 #define GET_ERRNO_H 1
-int get_errno();
+uint64_t get_errno();
 #endif
